+++ go/libraries/doltcore/dtestutils/data.go
@@ -15,21 +15,9 @@
 package dtestutils
 
 import (
-	"github.com/google/uuid"
-
-	"github.com/dolthub/dolt/go/libraries/doltcore/row"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	"github.com/dolthub/dolt/go/store/types"
 )
-
-var uuids = []uuid.UUID{
-	uuid.Must(uuid.Parse("00000000-0000-0000-0000-000000000000")),
-	uuid.Must(uuid.Parse("00000000-0000-0000-0000-000000000001")),
-	uuid.Must(uuid.Parse("00000000-0000-0000-0000-000000000002"))}
-var names = []string{"Bill Billerson", "John Johnson", "Rob Robertson"}
-var ages = []uint64{32, 25, 21}
-var titles = []string{"Senior Dufus", "Dufus", ""}
-var maritalStatus = []bool{true, false, false}
 
 const (
 	IdTag uint64 = iota
@@ -66,37 +54,4 @@ func Schema() (schema.Schema, error) {
 	}
 
 	return sch, err
-}
-
-// RowsAndSchema returns the schema and rows for the `people` test table.
-func RowsAndSchema() ([]row.Row, schema.Schema, error) {
-	sch, err := Schema()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	rows := make([]row.Row, len(uuids))
-	for i := 0; i < len(uuids); i++ {
-		married := types.Int(0)
-		if maritalStatus[i] {
-			married = types.Int(1)
-		}
-		taggedVals := row.TaggedValues{
-			IdTag:        types.String(uuids[i].String()),
-			NameTag:      types.String(names[i]),
-			AgeTag:       types.Uint(ages[i]),
-			TitleTag:     types.String(titles[i]),
-			IsMarriedTag: married,
-		}
-
-		r, err := row.New(types.Format_Default, sch, taggedVals)
-
-		if err != nil {
-			panic(err)
-		}
-
-		rows[i] = r
-	}
-
-	return rows, sch, err
 }
