+++ go/libraries/doltcore/dtestutils/testcommands/multienv.go
@@ -31,7 +31,6 @@ import (
 	"github.com/dolthub/dolt/go/libraries/doltcore/env/actions"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/dprocedures"
-	"github.com/dolthub/dolt/go/libraries/doltcore/table"
 	"github.com/dolthub/dolt/go/libraries/utils/config"
 	"github.com/dolthub/dolt/go/libraries/utils/filesys"
 	"github.com/dolthub/dolt/go/store/datas"
@@ -295,28 +294,14 @@ func (mr *MultiRepoTestSetup) CommitWithWorkingSet(dbName string) *doltdb.Commit
 	return commit
 }
 
-func createTestDataTable(ctx context.Context, ddb *doltdb.DoltDB) (*table.InMemTable, schema.Schema) {
-	rows, sch, err := dtestutils.RowsAndSchema()
-	if err != nil {
-		panic(err)
-	}
-
-	imt := table.NewInMemTable(sch)
-
-	for _, r := range rows {
-		err := imt.AppendRow(ctx, ddb.ValueReadWriter(), r)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	return imt, sch
-}
-
 func (mr *MultiRepoTestSetup) CreateTable(ctx context.Context, dbName, tblName string) {
 	dEnv := mr.envs[dbName]
 
-	_, sch := createTestDataTable(ctx, dEnv.DoltDB(ctx))
+	sch, err := dtestutils.Schema()
+	if err != nil {
+		mr.Errhand(err)
+	}
+
 	if err := createTestTable(dEnv, tblName, sch); err != nil {
 		mr.Errhand(err)
 	}
